# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /receptionist:
    post:
      operationId: createReceptionist
      tags: [receptionist]
      summary: Create a receptionist for the signed-in owner or legacy pilot
      parameters:
        - $ref: "#/components/parameters/SetupKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceptionistInput"
      responses:
        "201":
          description: Receptionist created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceptionistCreated"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Sign-in required or invalid pilot setup key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Pilot receptionist already provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Pilot setup key is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /receptionist/{receptionistId}/settings:
    get:
      operationId: getReceptionistSettings
      tags: [receptionist]
      summary: Get receptionist settings
      parameters:
        - $ref: "#/components/parameters/ReceptionistId"
        - $ref: "#/components/parameters/OwnerKey"
      responses:
        "200":
          description: Settings retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceptionistSettings"
        "401":
          description: Invalid owner key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Receptionist not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateReceptionistSettings
      tags: [receptionist]
      summary: Update receptionist settings
      parameters:
        - $ref: "#/components/parameters/ReceptionistId"
        - $ref: "#/components/parameters/OwnerKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceptionistSettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceptionistSettings"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid owner key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Receptionist not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /receptionist/{receptionistId}/conversations:
    get:
      operationId: listReceptionistConversations
      tags: [receptionist]
      summary: List conversations for this receptionist
      parameters:
        - $ref: "#/components/parameters/ReceptionistId"
        - $ref: "#/components/parameters/OwnerKey"
      responses:
        "200":
          description: Conversation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationPage"
        "401":
          description: Invalid owner key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Receptionist not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /receptionist/{receptionistId}/conversations/{conversationId}:
    get:
      operationId: getConversation
      tags: [receptionist]
      summary: Get a conversation with its messages
      parameters:
        - $ref: "#/components/parameters/ReceptionistId"
        - $ref: "#/components/parameters/ConversationId"
        - $ref: "#/components/parameters/OwnerKey"
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
        "401":
          description: Invalid owner key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /receptionist/{receptionistId}/conversations/{conversationId}/reply:
    post:
      operationId: ownerReply
      tags: [receptionist]
      summary: Send an owner reply into a conversation
      parameters:
        - $ref: "#/components/parameters/ReceptionistId"
        - $ref: "#/components/parameters/ConversationId"
        - $ref: "#/components/parameters/OwnerKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerReplyInput"
      responses:
        "201":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid owner key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /receptionist/{receptionistId}/conversations/{conversationId}/status:
    patch:
      operationId: updateConversationStatus
      tags: [receptionist]
      summary: Resolve or reopen a conversation
      parameters:
        - $ref: "#/components/parameters/ReceptionistId"
        - $ref: "#/components/parameters/ConversationId"
        - $ref: "#/components/parameters/OwnerKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationStatusUpdate"
      responses:
        "200":
          description: Conversation status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationStatus"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid owner key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conversation status changed before this update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    SetupKey:
      in: header
      name: X-SiteForge-Setup-Key
      required: false
      description: Required only for legacy pilot provisioning without a signed-in Clerk session
      schema:
        type: string
        minLength: 16
        maxLength: 256
...
    OwnerKey:
      in: header
      name: X-SiteForge-Owner-Key
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    ReceptionistId:
      in: path
      name: receptionistId
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    ConversationId:
      in: path
      name: conversationId
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required:
        - error

    # ─── Receptionist ────────────────────────────────────────────────────────
    ReceptionistInput:
      type: object
      additionalProperties: false
      properties:
        businessName:
          type: string
          minLength: 1
          maxLength: 200
        businessEmail:
          type: string
          maxLength: 200
        assistantName:
          type: string
          maxLength: 100
      required:
        - businessName
    ReceptionistCreated:
      type: object
      additionalProperties: false
      properties:
        receptionistId:
          type: string
        ownerKey:
          type: string
          description: One-time secret; store it — it will not be shown again
      required:
        - receptionistId
        - ownerKey
    ReceptionistSettings:
      type: object
      additionalProperties: false
      properties:
        receptionistId:
          type: string
        enabled:
          type: boolean
        businessName:
          type: string
        businessEmail:
          type: ["string", "null"]
        assistantName:
          type: string
        greeting:
          type: ["string", "null"]
        knowledge:
          type: ["string", "null"]
        faqs:
          type: ["string", "null"]
        hours:
          type: ["string", "null"]
        serviceArea:
          type: ["string", "null"]
        escalationContact:
          type: ["string", "null"]
        prohibitedActions:
          type: ["string", "null"]
        safeAutoReplyCategories:
          type: ["string", "null"]
        retellAgentId:
          type: ["string", "null"]
        retellPhoneNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - receptionistId
        - enabled
        - businessName
        - assistantName
        - createdAt
        - updatedAt
    ReceptionistSettingsUpdate:
      type: object
      additionalProperties: false
      properties:
        enabled:
          type: boolean
        businessName:
          type: string
          minLength: 1
          maxLength: 200
        businessEmail:
          type: ["string", "null"]
          maxLength: 200
        assistantName:
          type: string
          maxLength: 100
        greeting:
          type: ["string", "null"]
          maxLength: 2000
        knowledge:
          type: ["string", "null"]
          maxLength: 20000
        faqs:
          type: ["string", "null"]
          maxLength: 10000
        hours:
          type: ["string", "null"]
          maxLength: 2000
        serviceArea:
          type: ["string", "null"]
          maxLength: 1000
        escalationContact:
          type: ["string", "null"]
          maxLength: 500
        prohibitedActions:
          type: ["string", "null"]
          maxLength: 2000
        safeAutoReplyCategories:
          type: ["string", "null"]
          maxLength: 2000
        retellAgentId:
          type: ["string", "null"]
          maxLength: 200
        retellPhoneNumber:
          type: ["string", "null"]
          maxLength: 30

    # ─── Conversations & Messages ────────────────────────────────────────────
    ConversationSummary:
      type: object
      additionalProperties: false
      properties:
        conversationId:
          type: string
        receptionistId:
          type: string
        channel:
          type: string
        externalId:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        status:
          type: string
        messageCount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - conversationId
        - receptionistId
        - channel
        - status
        - messageCount
        - createdAt
        - updatedAt
    ConversationPage:
      type: object
      additionalProperties: false
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/ConversationSummary"
        total:
          type: number
      required:
        - conversations
        - total
    ChatMessage:
      type: object
      additionalProperties: false
      properties:
        messageId:
          type: string
        conversationId:
          type: string
        receptionistId:
          type: string
        role:
          type: string
        content:
          type: string
        approvalState:
          type: ["string", "null"]
        externalMessageId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - messageId
        - conversationId
        - receptionistId
        - role
        - content
        - createdAt
    AuditEvent:
      type: object
      additionalProperties: false
      properties:
        auditEventId:
          type: string
        eventType:
          type: string
        messageId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - auditEventId
        - eventType
        - createdAt
    ConversationDetail:
      type: object
      additionalProperties: false
      properties:
        conversationId:
          type: string
        receptionistId:
          type: string
        channel:
          type: string
        externalId:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        status:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        auditEvents:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - conversationId
        - receptionistId
        - channel
        - status
        - messages
        - auditEvents
        - createdAt
        - updatedAt
    OwnerReplyInput:
      type: object
      additionalProperties: false
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 10000
      required:
        - content
    ConversationStatusUpdate:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [open, closed]
        expectedUpdatedAt:
          type: string
          format: date-time
      required:
        - status
        - expectedUpdatedAt
    ConversationStatus:
      type: object
      additionalProperties: false
      properties:
        conversationId:
          type: string
        receptionistId:
          type: string
        status:
          type: string
          enum: [open, closed, escalated]
        updatedAt:
          type: string
          format: date-time
      required:
        - conversationId
        - receptionistId
        - status
        - updatedAt

    # ─── Gmail ───────────────────────────────────────────────────────────────
...